Object Explorer: Initial handling of object's attributes edition
