+++ spyder/plugins/variableexplorer/widgets/collectionseditor.py
@@ -28,19 +28,18 @@ import warnings
 # Third party imports
 import cloudpickle
 from qtpy.compat import getsavefilename, to_qvariant
-from qtpy.QtCore import (QAbstractTableModel, QDateTime, QModelIndex, Qt,
+from qtpy.QtCore import (QAbstractTableModel, QModelIndex, Qt,
                          Signal, Slot)
 from qtpy.QtGui import QColor, QKeySequence
-from qtpy.QtWidgets import (QAbstractItemDelegate, QApplication, QDateEdit,
-                            QDateTimeEdit, QDialog, QHBoxLayout, QHeaderView,
-                            QInputDialog, QItemDelegate, QLineEdit, QMenu,
+from qtpy.QtWidgets import (QApplication, QDialog, QHBoxLayout, QHeaderView,
+                            QInputDialog, QLineEdit, QMenu,
                             QMessageBox, QPushButton, QTableView,
                             QVBoxLayout, QWidget)
 from spyder_kernels.utils.misc import fix_reference_name
 from spyder_kernels.utils.nsview import (
-    array, DataFrame, Index, display_to_value, FakeObject,
-    get_color_name, get_human_readable_type, get_size, Image, is_editable_type,
-    is_known_type, MaskedArray, ndarray, np_savetxt, Series, sort_against,
+    DataFrame, display_to_value, FakeObject,
+    get_color_name, get_human_readable_type, get_size, Image,
+    MaskedArray, ndarray, np_savetxt, Series, sort_against,
     try_to_eval, unsorted_unique, value_to_display, get_object_attrs,
     get_type_string)
 
@@ -48,24 +47,15 @@ from spyder_kernels.utils.nsview import (
 from spyder.config.base import _, PICKLE_PROTOCOL
 from spyder.config.fonts import DEFAULT_SMALL_DELTA
 from spyder.config.gui import get_font
-from spyder.py3compat import (io, is_binary_string, is_text_string,
-                              PY3, to_text_string, is_type_text_string)
+from spyder.py3compat import (io, is_binary_string, PY3, to_text_string,
+                              is_type_text_string)
 from spyder.utils import icon_manager as ima
 from spyder.utils.misc import getcwd_or_home
 from spyder.utils.qthelpers import (add_actions, create_action,
                                     mimedata2url)
 from spyder.plugins.variableexplorer.widgets.importwizard import ImportWizard
-from spyder.plugins.variableexplorer.widgets.texteditor import TextEditor
-from spyder.plugins.variableexplorer.widgets.objectexplorer import (
-    ObjectExplorer)
-
-if ndarray is not FakeObject:
-    from spyder.plugins.variableexplorer.widgets.arrayeditor import (
-            ArrayEditor)
-
-if DataFrame is not FakeObject:
-    from spyder.plugins.variableexplorer.widgets.dataframeeditor import (
-            DataFrameEditor)
+from spyder.plugins.variableexplorer.widgets.collectionsdelegate import (
+    CollectionsDelegate)
 
 
 # Maximum length of a serialized variable to be set in the kernel
@@ -421,294 +411,6 @@ class CollectionsModel(ReadOnlyCollectionsModel):
         self.set_value(index, value)
         self.dataChanged.emit(index, index)
         return True
-
-
-class CollectionsDelegate(QItemDelegate):
-    """CollectionsEditor Item Delegate"""
-    sig_free_memory = Signal()
-
-    def __init__(self, parent=None):
-        QItemDelegate.__init__(self, parent)
-        self._editors = {} # keep references on opened editors
-        
-    def get_value(self, index):
-        if index.isValid():
-            return index.model().get_value(index)
-    
-    def set_value(self, index, value):
-        if index.isValid():
-            index.model().set_value(index, value)
-
-    def show_warning(self, index):
-        """
-        Decide if showing a warning when the user is trying to view
-        a big variable associated to a Tablemodel index
-
-        This avoids getting the variables' value to know its
-        size and type, using instead those already computed by
-        the TableModel.
-        
-        The problem is when a variable is too big, it can take a
-        lot of time just to get its value
-        """
-        try:
-            val_size = index.model().sizes[index.row()]
-            val_type = index.model().types[index.row()]
-        except:
-            return False
-        if val_type in ['list', 'set', 'tuple', 'dict'] and \
-                int(val_size) > 1e5:
-            return True
-        else:
-            return False
-
-    def createEditor(self, parent, option, index, object_explorer=False):
-        """Overriding method createEditor"""
-        if index.column() < 3:
-            return None
-        if self.show_warning(index):
-            answer = QMessageBox.warning(self.parent(), _("Warning"),
-                                      _("Opening this variable can be slow\n\n"
-                                        "Do you want to continue anyway?"),
-                                      QMessageBox.Yes | QMessageBox.No)
-            if answer == QMessageBox.No:
-                return None
-        try:
-            value = self.get_value(index)
-            if value is None:
-                return None
-        except Exception as msg:
-            QMessageBox.critical(self.parent(), _("Error"),
-                                 _("Spyder was unable to retrieve the value of "
-                                   "this variable from the console.<br><br>"
-                                   "The error mesage was:<br>"
-                                   "<i>%s</i>"
-                                   ) % to_text_string(msg))
-            return
-        key = index.model().get_key(index)
-        readonly = (isinstance(value, (tuple, set)) or self.parent().readonly
-                    or not is_known_type(value))
-        # CollectionsEditor for a list, tuple, dict, etc.
-        if isinstance(value, (list, set, tuple, dict)) and not object_explorer:
-            editor = CollectionsEditor(parent=parent)
-            editor.setup(value, key, icon=self.parent().windowIcon(),
-                         readonly=readonly)
-            self.create_dialog(editor, dict(model=index.model(), editor=editor,
-                                            key=key, readonly=readonly))
-            return None
-        # ArrayEditor for a Numpy array
-        elif (isinstance(value, (ndarray, MaskedArray)) and
-                ndarray is not FakeObject and not object_explorer):
-            editor = ArrayEditor(parent=parent)
-            if not editor.setup_and_check(value, title=key, readonly=readonly):
-                return
-            self.create_dialog(editor, dict(model=index.model(), editor=editor,
-                                            key=key, readonly=readonly))
-            return None
-        # ArrayEditor for an images
-        elif (isinstance(value, Image) and ndarray is not FakeObject and
-                Image is not FakeObject and not object_explorer):
-            arr = array(value)
-            editor = ArrayEditor(parent=parent)
-            if not editor.setup_and_check(arr, title=key, readonly=readonly):
-                return
-            conv_func = lambda arr: Image.fromarray(arr, mode=value.mode)
-            self.create_dialog(editor, dict(model=index.model(), editor=editor,
-                                            key=key, readonly=readonly,
-                                            conv=conv_func))
-            return None
-        # DataFrameEditor for a pandas dataframe, series or index
-        elif (isinstance(value, (DataFrame, Index, Series))
-                and DataFrame is not FakeObject and not object_explorer):
-            editor = DataFrameEditor(parent=parent)
-            if not editor.setup_and_check(value, title=key):
-                return
-            editor.dataModel.set_format(index.model().dataframe_format)
-            editor.sig_option_changed.connect(self.change_option)
-            self.create_dialog(editor, dict(model=index.model(), editor=editor,
-                                            key=key, readonly=readonly))
-            return None
-        # QDateEdit and QDateTimeEdit for a dates or datetime respectively
-        elif isinstance(value, datetime.date) and not object_explorer:
-            if readonly:
-                return None
-            else:
-                if isinstance(value, datetime.datetime):
-                    editor = QDateTimeEdit(value, parent=parent)
-                else:
-                    editor = QDateEdit(value, parent=parent)
-                editor.setCalendarPopup(True)
-                editor.setFont(get_font(font_size_delta=DEFAULT_SMALL_DELTA))
-                return editor
-        # TextEditor for a long string
-        elif is_text_string(value) and len(value) > 40 and not object_explorer:
-            te = TextEditor(None, parent=parent)
-            if te.setup_and_check(value):
-                editor = TextEditor(value, key,
-                                    readonly=readonly, parent=parent)
-                self.create_dialog(editor, dict(model=index.model(),
-                                                editor=editor, key=key,
-                                                readonly=readonly))
-            return None
-        # QLineEdit for an individual value (int, float, short string, etc)
-        elif is_editable_type(value) and not object_explorer:
-            if readonly:
-                return None
-            else:
-                editor = QLineEdit(parent=parent)
-                editor.setFont(get_font(font_size_delta=DEFAULT_SMALL_DELTA))
-                editor.setAlignment(Qt.AlignLeft)
-                # This is making Spyder crash because the QLineEdit that it's
-                # been modified is removed and a new one is created after
-                # evaluation. So the object on which this method is trying to
-                # act doesn't exist anymore.
-                # editor.returnPressed.connect(self.commitAndCloseEditor)
-                return editor
-        # ObjectExplorer for an arbitrary Python object
-        else:
-            show_callable_attributes = index.model().show_callable_attributes
-            show_special_attributes = index.model().show_special_attributes
-
-            if show_callable_attributes is None:
-                show_callable_attributes = False
-            if show_special_attributes is None:
-                show_special_attributes = False
-
-            key = index.model().keys[index.row()]
-            editor = ObjectExplorer(
-                value,
-                name=key,
-                parent=parent,
-                show_callable_attributes=show_callable_attributes,
-                show_special_attributes=show_special_attributes)
-            editor.sig_option_changed.connect(self.change_option)
-            self.create_dialog(editor, dict(model=index.model(),
-                                            editor=editor,
-                                            key=key, readonly=readonly))
-            return None
-
-    def create_dialog(self, editor, data):
-        self._editors[id(editor)] = data
-        editor.accepted.connect(
-                     lambda eid=id(editor): self.editor_accepted(eid))
-        editor.rejected.connect(
-                     lambda eid=id(editor): self.editor_rejected(eid))
-        editor.show()
-
-    @Slot(str, object)
-    def change_option(self, option_name, new_value):
-        """
-        Change configuration option.
-
-        This function is called when a `sig_option_changed` signal is received.
-        At the moment, this signal can only come from a DataFrameEditor
-        or an ObjectExplorer.
-        """
-        if option_name == 'dataframe_format':
-            self.parent().set_dataframe_format(new_value)
-        elif option_name == 'show_callable_attributes':
-            self.parent().toggle_show_callable_attributes(new_value)
-        elif option_name == 'show_special_attributes':
-            self.parent().toggle_show_special_attributes(new_value)
-
-    def editor_accepted(self, editor_id):
-        data = self._editors[editor_id]
-        if not data['readonly']:
-            index = data['model'].get_index_from_key(data['key'])
-            value = data['editor'].get_value()
-            conv_func = data.get('conv', lambda v: v)
-            self.set_value(index, conv_func(value))
-        # This is needed to avoid the problem reported on
-        # issue 8557
-        try:
-            self._editors.pop(editor_id)
-        except KeyError:
-            pass
-        self.free_memory()
-        
-    def editor_rejected(self, editor_id):
-        # This is needed to avoid the problem reported on
-        # issue 8557
-        try:
-            self._editors.pop(editor_id)
-        except KeyError:
-            pass
-        self.free_memory()
-
-    def free_memory(self):
-        """Free memory after closing an editor."""
-        try:
-            self.sig_free_memory.emit()
-        except RuntimeError:
-            pass
-
-    def commitAndCloseEditor(self):
-        """Overriding method commitAndCloseEditor"""
-        editor = self.sender()
-        # Avoid a segfault with PyQt5. Variable value won't be changed
-        # but at least Spyder won't crash. It seems generated by a bug in sip.
-        try:
-            self.commitData.emit(editor)
-        except AttributeError:
-            pass
-        self.closeEditor.emit(editor, QAbstractItemDelegate.NoHint)
-
-    def setEditorData(self, editor, index):
-        """
-        Overriding method setEditorData
-        Model --> Editor
-        """
-        value = self.get_value(index)
-        if isinstance(editor, QLineEdit):
-            if is_binary_string(value):
-                try:
-                    value = to_text_string(value, 'utf8')
-                except:
-                    pass
-            if not is_text_string(value):
-                value = repr(value)
-            editor.setText(value)
-        elif isinstance(editor, QDateEdit):
-            editor.setDate(value)
-        elif isinstance(editor, QDateTimeEdit):
-            editor.setDateTime(QDateTime(value.date(), value.time()))
-
-    def setModelData(self, editor, model, index):
-        """
-        Overriding method setModelData
-        Editor --> Model
-        """
-        if not hasattr(model, "set_value"):
-            # Read-only mode
-            return
-        
-        if isinstance(editor, QLineEdit):
-            value = editor.text()
-            try:
-                value = display_to_value(to_qvariant(value),
-                                         self.get_value(index),
-                                         ignore_errors=False)
-            except Exception as msg:
-                raise
-                QMessageBox.critical(editor, _("Edit item"),
-                                     _("<b>Unable to assign data to item.</b>"
-                                       "<br><br>Error message:<br>%s"
-                                       ) % str(msg))
-                return
-        elif isinstance(editor, QDateEdit):
-            qdate = editor.date()
-            value = datetime.date( qdate.year(), qdate.month(), qdate.day() )
-        elif isinstance(editor, QDateTimeEdit):
-            qdatetime = editor.dateTime()
-            qdate = qdatetime.date()
-            qtime = qdatetime.time()
-            value = datetime.datetime( qdate.year(), qdate.month(),
-                                       qdate.day(), qtime.hour(),
-                                       qtime.minute(), qtime.second() )
-        else:
-            # Should not happen...
-            raise RuntimeError("Unsupported editor widget")
-        self.set_value(index, value)
 
 
 class BaseHeaderView(QHeaderView):
